Create Sal's Shipping from Code Academy
I didn't realize until after finishing that I had completely overthought this program. I'm not sure if that's a good thing or a bad thing. The sample program was way simpler. I wouldn't mind some constructive feedback on whether I did a lot of needless work and that is a problem in programming. Or if I was thorough and covered my bases and that's find. 
Either way, I'm on day 2, so it's all a learning process