Add files via upload
As soon as I got to this lesson in Code Academy with all the names listed in the votes variable, I wanted to see write this program to see if I could. It's super simple way to figure out who got the most votes. I am already thinking of the lists I could make that can help me with my PCs as a D&D player and so many other things as a Game Master.